Use a queue to send SMS
