+++ .gitlab-ci.yml
@@ -56,6 +56,7 @@ deploy_review:
       --set imageCredentials.password="${DOCKER_REGISTRY_PASSWORD}"
       --set imageCredentials.registry="${CI_REGISTRY}"
       --set env.REDIS="${STAGING_REDIS}"
+      --set env.AMQP="${STAGING_AMQP}"
       --set env.MONGODB="${STAGING_MONGODB}"
       --set env.TWILIO__SID="${STAGING_TWILIO__SID}"
       --set env.TWILIO__TOKEN="${STAGING_TWILIO__TOKEN}"
@@ -98,6 +99,7 @@ deploy_staging:
       --set imageCredentials.password="${DOCKER_REGISTRY_PASSWORD}"
       --set imageCredentials.registry="${CI_REGISTRY}"
       --set env.REDIS="${STAGING_REDIS}"
+      --set env.AMQP="${STAGING_AMQP}"
       --set env.MONGODB="${STAGING_MONGODB}"
       --set env.TWILIO__SID="${STAGING_TWILIO__SID}"
       --set env.TWILIO__TOKEN="${STAGING_TWILIO__TOKEN}"
@@ -142,6 +144,7 @@ deploy_prod:
       --set imageCredentials.password="${DOCKER_REGISTRY_PASSWORD}"
       --set imageCredentials.registry="${CI_REGISTRY}"
       --set env.REDIS="${PRODUCTION_REDIS}"
+      --set env.AMQP="${PRODUCTION_AMQP}"
       --set env.MONGODB="${PRODUCTION_MONGODB}"
       --set env.TWILIO__SID="${PRODUCTION_TWILIO__SID}"
       --set env.TWILIO__TOKEN="${PRODUCTION_TWILIO__TOKEN}"
