+++ package.json
@@ -6,11 +6,12 @@
     "test": "mocha --experimental-modules --exit --require @std/esm --recursive \"tests/**/*.mjs\"",
     "serve": "LOG=database node --experimental-modules --es-module-specifier-resolution=node server.mjs",
     "nodemon": "nodemon --experimental-modules --es-module-specifier-resolution=node --inspect server.mjs",
-    "lint": "eslint --ext .js,.mjs --ignore-path .gitignore .",
+    "lint": "eslint --ext .js,.mjs --ignore-path .gitignore --ignore-path .eslintignore .",
     "generate-vapid": "node generate-vapid-key.js",
     "precommit": "npm run lint",
     "generate-state-schema": "node --experimental-modules --es-module-specifier-resolution=node tests/generate-state-schema.js",
-    "import-fixtures": "node --experimental-modules --es-module-specifier-resolution=node fixtures"
+    "import-fixtures": "node --experimental-modules --es-module-specifier-resolution=node fixtures",
+    "consumer:send-sms": "node --experimental-modules --es-module-specifier-resolution=node consumers/send-sms.js"
   },
   "author": "Gaël",
   "license": "ISC",
@@ -54,8 +55,8 @@
     "nconf": "^0.10.0",
     "node-input-validator": "^3.8.0",
     "nodemailer": "^6.3.1",
-    "nodemon": "^1.19.4",
     "qs": "^6.9.1",
+    "rhea": "^1.0.15",
     "socket.io": "^2.3.0",
     "socket.io-redis": "^5.2.0",
     "sprintf-lite": "^1.3.0",
@@ -75,6 +76,7 @@
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-plugin-import": "^2.18.2",
     "mocha": "^6.2.2",
+    "nodemon": "^2.0.2",
     "supertest": "^4.0.2"
   },
   "peerDependencies": {
